# server/src/openapi/API.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Somni joyas",
        "description": "API for SomnoJoyas DB management"
    },
    "servers": [
        {
            "url": "http://localhost:3000"
        },
        {
            "url": "https://somniapi.onrender.com"
        }
    ],
    "components": {
        "schemas": {
...
            "Article": {
                "type": "object",
                "properties": {
                    "type": {
                        "type": "string",
                        "enum": [
                            "anillo",
                            "pendientes",
                            "pulseras",
                            "cadenas",
                            "gargantillas",
                            "colgante"
                        ]
                    },
                    "material": {
                        "type": "string",
                        "enum": [
                            "plata",
                            "oro",
                            "acero"
                        ]
                    },
                    "finish": {
                        "type": "string",
                        "enum": [
                            "plateado",
                            "dorado",
                            "oro rosado",
                            "rodiado"
                        ]
                    },
                    "dimensions": {
                        "type": "string"
                    },
                    "details": {
                        "type": "string"
                    },
                    "units": {
                        "type": "number"
                    },
                    "price": {
                        "type": "number"
                    },
                    "image": {
                        "type": "string"
                    }
                },
                "required": [
                    "type",
                    "material",
                    "finish",
                    "details",
                    "units",
                    "price",
                    "image"
                ],
                "example": {
                    "type": "anillo",
                    "material": "plata",
                    "finish": "plateado",
                    "dimensions": "10mm",
                    "details": "Anillo de plata con diamantes incrustados.",
                    "units": 10,
                    "price": 100,
                    "image": "https://example.com/ring.jpg"
                }
            },
...
        },
        "securitySchemes": {
            "securitySchemes": {
                "JWTAuth": {
                    "type": "http",
                    "scheme": "bearer",
                    "bearerFormat": "JWT"
                }
            }
        },
...
    },
    "paths": {
...
        "/articles": {
            "get": {
                "tags": [
                    "Articles"
                ],
                "summary": "Get all articles",
                "operationId": "getArticles",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                }
            },
            "post": {
                "tags": [
                    "Articles"
                ],
                "summary": "Create a new article",
                "operationId": "createArticle",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Article"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Article created successfully"
                    }
                }
            }
        },
        "/articles/{id}": {
            "get": {
                "tags": [
                    "Articles"
                ],
                "summary": "Get article by ID",
                "operationId": "getArticleById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "description": "ID of the article to get",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                }
            },
            "put": {
                "tags": [
                    "Articles"
                ],
                "summary": "Update article by ID",
                "operationId": "updateArticleById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "description": "ID of the article to update",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Article"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Article updated successfully"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Articles"
                ],
                "summary": "Delete article by ID",
                "operationId": "deleteArticleById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "description": "ID of the article to delete",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Article deleted successfully"
                    }
                }
            }
        },
...
    },
...
}